crate_release.md: update documentation
This commit unifies the steps required for publishing crates that have
their own repositories with ones that are part of a workspace, such that
it's easier to follow the process. Another proposed change to make things
uniform is to always include the name of the crate as part of tag and
release names (i.e. `crate-name-v1.2.3` vs just `v1.2.3`).

Signed-off-by: Alexandru Agache <aagch@amazon.com>